refactor: ViewModel 처리 형태로 변경
