+++ WeatherForecast/WeatherForecast/SceneDelegate.swift
@@ -19,7 +19,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
         let locationManager = CLLocationManager(desiredAccuracy: kCLLocationAccuracyThreeKilometers)
         let locationService = LocationService(locationManager: locationManager)
-        let mainViewController = MainViewController(locationService: locationService)
+        let weatherDataViewModel = WeatherDataViewModel(locationService: locationService)
+        let mainViewController = MainViewController(weatherDataViewModel: weatherDataViewModel)
         
         window?.rootViewController = mainViewController
         window?.makeKeyAndVisible()
